+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Login-code checks the given login field instead of email.
 - It is now possible to add your own translation directly in the plugin. (Again)
 - Added extra filter to course list (ajax) to skip "next" event if there isn't a public price name set.
+- Fixes [#50](https://github.com/MultinetInteractive/EduAdmin-WordPress/issues/50), so end users won't have to fill in their details three times.
 
 ## [0.9.9.2.40]
 ### Added